Circular Queue generate by Array based program
Write a program to perform the following operations using Java.
1. Read the count of elements from user to create the circular queue using array. 
2. Create the class and methods for the following and call them using a menu based program and 
call them using a loop.
a. Show the contents of circular queue
b. Enqueue the data in the circular queue
c. Access the data from the circular queue
d. Check how many elements are there in the circular queue.
e. Check to see if the circular queue is full?
f. Check to see if circular queue is empty.